manager view/low view added, add stock WIP

+++ SQLrunazon.sql
@@ -8,7 +8,7 @@ CREATE TABLE inventory (
 item_id INT NOT NULL AUTO_INCREMENT,
 product_name VARCHAR(20) NOT NULL,
 department_name VARCHAR(20),
-price DECIMAL(5,2) NOT NULL,
+price INT(3) NOT NULL,
 stock_quantity INT(2) NOT NULL,
 PRIMARY KEY (item_id)
 );
